refactor(output): extract comment tests to sibling file
